Moved DAO and DaoPair to the interfaces directory.
